extract types to type folder
