+++ frontend/app/provider/page-container.tsx
@@ -2,7 +2,7 @@ import { useGetAllBouts } from "@/hooks/use-get-all-bouts";
 import { useSuspenseGetSyncData } from "@/hooks/use-suspense-get-sync-data";
 import { SyncDataProvider } from "@/hooks/use-sync-context";
 import queryClient from "@/lib/cache";
-import { Bout } from "@/lib/game/bouts";
+import { Bout } from "@/types/bout";
 import { MantineProvider } from "@mantine/core";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { PropsWithChildren, StrictMode, useEffect, useState } from "react";
