+++ frontend/hooks/use-get-all-bouts.ts
@@ -1,6 +1,6 @@
 import queryClient from "@/lib/cache";
-import { Bout } from "@/lib/game/bouts";
 import { localAPI } from "@/lib/requests";
+import { Bout } from "@/types/bout";
 import { AppQueryOptions } from "@/types/query";
 import { useQuery } from "@tanstack/react-query";
 import { useCallback } from "react";
